requirements.txt
flask==3.0.3
python-dotenv==1.0.1
requests==2.32.3
PyQt6==6.7.0
PyQt6-WebEngine==6.7.0